+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Monitoring namespace based on hostname, process id and thread id ([#1222](https://github.com/neptune-ai/neptune-client/pull/1222))
 - Removed deprecated `--run` option from `neptune sync` command ([#1231](https://github.com/neptune-ai/neptune-client/pull/1231))
 - Removed `Run._short_id` property ([#1234](https://github.com/neptune-ai/neptune-client/pull/1234))
+- Removed `get_run_url` method ([#1238](https://github.com/neptune-ai/neptune-client/pull/1238))
 
 ## neptune-client 0.16.18